+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Version - 0]
+## improve [v0.34.6] 2023-06-19
+- enhance news menu
+- added category news & challenge
 ## improve [v0.34.5] 2023-06-16
 - enhanced quota report
 ## improve [v0.34.4] 2023-06-16